+++ auto_calender.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+import re
 from option import options
 from bs4 import BeautifulSoup
 from datetime import datetime, timedelta
@@ -8,7 +9,6 @@ def init():
 	for i in range(5):
 		for j in range(11):
 			delta.append(i+1)
-	print(len(delta))
 	return delta
 
 def get_time(time, delta):
@@ -26,8 +26,16 @@ def get_str(name, loc, start_time, end_time, delta):
 	str += f"SUMMARY:{name}\n"
 	str += f"LOCATION:{loc}\n"
 	str += 'END:VEVENT\n'
-	print(str)
+	# print(str)
 	return str
+
+def check(name):
+	if opt.exclude is False:
+		return False
+	for str in opt.exclude_class:
+		if name.count(str) > 0:
+			return True
+	return False
 
 if __name__ == '__main__':
 
@@ -36,11 +44,22 @@ if __name__ == '__main__':
 	print(opt)
 
 	with open(opt.read_path,"r") as f:
-		data=f.readline()
+		data=f.read()
 	
 	soup = BeautifulSoup(data, 'lxml')
 	divs = soup.find_all('div')
+	h2 = soup.find_all('h2')
 	# get all the data we need
+
+	for k in h2:
+		if k.string is not None:
+			date_list = re.findall(r"\d+", k.string)
+			# print(date_list, k.string)
+			opt.year = date_list[0]
+			opt.month = date_list[1]
+			opt.day = date_list[2]
+			# print(opt.year,opt.month,opt.day)
+			break
 
 	classes = []
 	for k in divs:
@@ -55,6 +74,8 @@ if __name__ == '__main__':
 		i = 2 * k
 		j = 2 * k + 1
 		name = classes[i][:-11]
+		if(check(name)):
+			continue;
 		time = classes[i][-11:]
 		time_start, time_end = time.split('-')
 		loc = classes[j]
