+++ option.py
@@ -14,8 +14,9 @@ class options():
         parser.add_argument("--month", type=int, default=9, help="The starting day of the week.")
         parser.add_argument("--year", type=int, default=2023, help="The starting day of the week.")
         parser.add_argument("--save_path", type=str, default='export.ics', help="The path to store the calender.")
-        parser.add_argument("--read_path", type=str, default="date.txt", help="The read path.")
+        parser.add_argument("--read_path", type=str, default="我的日程.html", help="The read path.")
         parser.add_argument("--head_path", type=str, default="header.txt", help="The first few lines of the document.")
+        parser.add_argument("--exclude", action="store_true", default=False, help="Whether to exclude some class or not.")
         parser.add_argument("--exclude_class", nargs='+', default=["自习","升旗"])
         self.opt = parser.parse_args()
         self.init()
